[REFACTOR] 3. Factor out Top-Main-Side Layout
